Fixing timestamp type in tests

+++ core/services/relay/evm/mercury/data_source_test.go
@@ -52,7 +52,7 @@ func TestMercurySetCurrentBlock(t *testing.T) {
 
 		assert.Equal(t, h.Number, obs.CurrentBlockNum.Val)
 		assert.Equal(t, h.Hash.Bytes(), obs.CurrentBlockHash.Val)
-		assert.Equal(t, h.Timestamp.Unix(), obs.CurrentBlockTimestamp.Val)
+		assert.Equal(t, uint64(h.Timestamp.Unix()), obs.CurrentBlockTimestamp.Val)
 
 		chainHeadTracker.AssertExpectations(t)
 		headTracker.AssertExpectations(t)
@@ -76,7 +76,7 @@ func TestMercurySetCurrentBlock(t *testing.T) {
 
 		assert.Equal(t, h.Number, obs.CurrentBlockNum.Val)
 		assert.Equal(t, h.Hash.Bytes(), obs.CurrentBlockHash.Val)
-		assert.Equal(t, h.Timestamp.Unix(), obs.CurrentBlockTimestamp.Val)
+		assert.Equal(t, uint64(h.Timestamp.Unix()), obs.CurrentBlockTimestamp.Val)
 
 		chainHeadTracker.AssertExpectations(t)
 		ethClient.AssertExpectations(t)
